Auto-save option: Add 'auto-save' preference and save onPause() that defaults to false.
If enabled, a note that is being edited will be saved as soon as
the app loses focus.

Also, if enabled, don't prompt while leaving EditNoteActivity
onBackPressed() - it's getting saved anyway onPause().

Note: a more (UX) elegant solution would be to save as a draft
(keeping the old note too) and to somehow show it in the UI.